+++ CHANGELOG.md
@@ -3,25 +3,36 @@
 ## [v0.1.2] - 2024-12-21
 
 ### Added
+
+- Toast when deleting products from cart
 - Mobile navbar collapse when click on cart and language
+- Cart overview section
 
 ### Changed
+
 - Optimized Cart items view on mobile
+- `Toast.tsx` now accepts success and danger variants
+- Extracted cart overview, functions and items into seperate files
+- Restructured keys in `translations.json`
 
 ### Removed
+
 - Sticky navbar
 
 ## [v0.1.1] - 2024-12-11
 
 ### Added
+
 - Banner to `CartIcon.tsx` to show the product count using global state
 - Display products in cart in `Cart.tsx` as a list extracted from session storage
 - Navbar collapse when a menu item is clicked on mobile
 
 ### Renamed
+
 - `PartnersParent.tsx` and `CatalogParent.tsx` into `...Page.tsx`
 
 ### Changed
+
 - Toasts to show in the center
 - Products get saved to local storage instead of session storage
 - Made the cart icon banner round
@@ -29,17 +40,16 @@
 - `Section.tsx` uses full width
 - Toasts stay for 3 seconds
 
-
-
 ### Moved
+
 - `Cart.tsx` under `cart` folder
 - - Styles from `Toast.tsx` into its css file
 - `CartIcon.tsx` styles in css file
 
 ### Fixed
+
 - Toast flashing after disappearing due to animation timing
 - Cart items dont fit on mobile
-
 
 ## [v0.1.0] - 2024-11-2